Move NumericInput to the right folder
